refactored the comparators
